SQL and user class update
SQL class has been updated to verify, validate, insert and delete users.
Functions for selecting, inserting and deleting loans, incomes, and
expenses are all done. I did as much testing as I reasonably could
without changing all kinds of code and everything appears to be in
working order in the mysqlconnect class. The current database is setup
to my localhost, I'll try to get that changed tomorrow now that Jose got
his remote database working. The hashtables are still present but we've
decided not to use them, at least at this point. We can use one of the
unique identifiers for identifying stuff in tables for the moment just
so we have the basic functionality and we can improve on that later if
need be. There's a few comments in the code but not many. Most of it
becomes pretty familiar after going through a few of these functions.

The user.java file was updated to only contain email and password as I
was told that would be all we would be using for this iteration and I
thank whoever made that decision. Made life easier