Debt Payoff Calculator Source in Progress
This is the test bed for the debt payoff calculator.  It is not ready
for a commit to the main project folder, so I'm uploading it here as a
jumping point for others to work off of.

To work on these files, create a new JavaFX project in eclipse.  Delete
the src folder and the Application Class folder (that is created
automatically by Eclipse).  Then copy the "src - DEBTPAYOFF TESTBED" to
the new JavaFX project folder and then finally rename the folder (now
stored locally on your system) simply to "src".  I added the caps to the
src folder name to help prevent accidentally overwriting the src folder
for the main app.

Finally, near the bottom of the code I have the original code that we
tried commented out.  I left it there for a reference point for now.