Debt Calculator
The debt calculator functionality is now working. It will return an
array list of every payment for every debt. The array list contains the
data points necessary for the graph(month and principle). There are two
classes that where added who's sole purpose is to generate the debt log
payment history.