Updated GUI for Snowball and Avalanche
I added a button for the snowball method in the GUI.  I also fixed some
of the naming conventions for the debt payoff handlers as well as the
actual calculations (I updated them to reflect whether they are for
snowball or avalanche).  I opted to create a separate handler for the
snowball method rather than use the same handler for both (that would
have used a conditional to figure out which calculation to execute based
on which button was pressed).  This was mainly from reading online that
it is generally bad practice to use the same handler for multiple
buttons, even if they execute very similar code.  It should be an easy
fix if we want to change it back to a single button handler.